+++ src/dashboard/ui/src/lib/Icon.svelte
@@ -217,6 +217,52 @@
 				}
 			}
 		],
+		download: [
+			{
+				tag: "path",
+				attrs: {
+					d: "M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4"
+				}
+			},
+			{
+				tag: "polyline",
+				attrs: {
+					points: "7 10 12 15 17 10"
+				}
+			},
+			{
+				tag: "line",
+				attrs: {
+					x1: "12",
+					x2: "12",
+					y1: "15",
+					y2: "3"
+				}
+			}
+		],
+		upload: [
+			{
+				tag: "path",
+				attrs: {
+					d: "M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4"
+				}
+			},
+			{
+				tag: "polyline",
+				attrs: {
+					points: "17 8 12 3 7 8"
+				}
+			},
+			{
+				tag: "line",
+				attrs: {
+					x1: "12",
+					x2: "12",
+					y1: "3",
+					y2: "15"
+				}
+			}
+		],
 		sun: [
 			{
 				tag: "circle",